requirements.txt
# Core ML/DL Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.55.0

# Document Processing
python-doctr[torch]>=1.0.0
Pillow>=9.2.0

# Data Processing
numpy>=1.16.0
pandas>=1.3.0

# Training and Evaluation
accelerate>=0.26.0
datasets>=2.14.0
scikit-learn>=1.0.0
seqeval>=1.2.2

# Visualization and Progress
tqdm>=4.30.0
matplotlib>=3.5.0

# Utilities
requests>=2.25.0
packaging>=20.9
